requirements.txt
python-ev3dev2
matplotlib
seaborn